requirements.txt
amqp==5.1.1
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
billiard==3.6.4.0
biopython==1.80
boto3==1.26.32
botocore==1.29.32
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-simple-history==3.3.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.2
numpy==1.24.0
packaging==23.0
prompt-toolkit==3.0.33
psycopg2==2.9.5
pycparser==2.21
pycurl==7.44.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2022.7.1
ratelimit==2.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.5
xmltodict==0.13.0
xmltramp2==3.1.1
